Improved build scripts: * better error detection * detecting msbuild and vstest location * web.config backup only when needed * CleanBuildTest.ps1 for testing full build without cached binaries

+++ tools/Build/Remove-RhetosServer.ps1
@@ -1,7 +1,8 @@
-param([Parameter(Mandatory=$true)][string]$destPath)
-
 # Delete old version of Rhetos server before installing a new one,
 # while preserving certain .config, .linq and Logs files.
+
+$ErrorActionPreference = 'Stop'
+$destPath = "$PSScriptRoot\..\..\dist\BookstoreRhetosServer"
 
 if (Test-Path "$destPath\bin")
   { Get-ChildItem "$destPath\bin" | Where-Object Name -ne 'ConnectionStrings.config' | Remove-Item -Recurse }
@@ -12,7 +13,19 @@ if (Test-Path "$destPath\PackagesCache")
 if (Test-Path "$destPath\Resources")
   { Remove-Item "$destPath\Resources" -Recurse }
 if (Test-Path "$destPath\Web.config")
-  { Copy-Item "$destPath\Web.config" ("$destPath\Web.config." + (Get-Date -Format "yyyyMMdd.HHmmss") + ".backup") }
+{
+  $sameBackupExists = $false;
+  if (Test-Path "$destPath\Web.config*.backup")
+  {
+    $lastBackup = Get-ChildItem "$destPath\Web.config*.backup" | Sort-Object Name | Select-Object -Last 1 -ExpandProperty FullName
+    $diff = Compare-Object -ReferenceObject $(Get-Content "$destPath\Web.config") -DifferenceObject $(Get-Content $lastBackup)
+    if ($diff.Length -eq 0)
+      { $sameBackupExists = $true }
+  }
+  if (!$sameBackupExists)
+    { Copy-Item "$destPath\Web.config" ("$destPath\Web.config." + (Get-Date -Format "yyyyMMdd.HHmmss") + ".backup") }
+}
+
 $rootFiles = 'ChangeLog.md','Default.aspx','Global.asax','Readme.md','Rhetos Server DOM.linq','Rhetos Server SOAP.linq','RhetosService.svc','Template.RhetosPackages.config','Template.RhetosPackageSources.config','Web.config'
 Get-ChildItem $destPath | Where-Object Name -In $rootFiles | Remove-Item
